Update: "cart & order에서 자잘한 오류나는 현상 해결"
cart & order에서 자잘한 오류나는 현상 해결

+++ src/main/java/com/mycompany/miniproject/controller/OrderController.java
@@ -9,7 +9,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -72,9 +71,9 @@ public class OrderController {
 	}
 	
 	@Secured("ROLE_USER")
-	@GetMapping("/cartAdd")
+	@PostMapping("/cartAdd")
 	public ResponseEntity<String> cartAdd(@RequestParam(defaultValue="1") int productQty, int productId, Model model, Authentication authentication) {
-		
+		log.info("실행");
 		HttpHeaders headers = new HttpHeaders();
 		headers.add("Content-Type", "text/html; charset=UTF-8");
 	    
@@ -106,10 +105,14 @@ public class OrderController {
 	}
 	
 	@GetMapping("/deleteProduct")
-	public String deleteProduct(int productId, Authentication authentication) {
+	public ResponseEntity<Integer> deleteProduct(int productId, Authentication authentication) {
 		String userId = authentication.getName();
-		cartService.deleteProduct(productId, userId);
-		return "/order/cart";
+		boolean hasProduct = cartService.hasProductInCart(productId, userId);
+		if(hasProduct) {
+			cartService.deleteProduct(productId, userId);
+			return ResponseEntity.ok(1);
+		}else
+			return ResponseEntity.ok(0);
 	}
 	
 	@RequestMapping("/order")
